requirements.txt
PyQt5
Pillow
pillow-heif

